# app/static/swagger.yaml
swagger: '2.0'
info:
  title: Mechanic Shop API
  description: REST API for customers, mechanics, inventory, and service tickets.
  version: 1.0.0
host: "mechanic-shop-api-4ig7.onrender.com"
basePath: /
schemes:
  - "https"
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'Use format: Bearer <token>'
paths:
...
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: Create a mechanic
      description: Creates a new mechanic record. Rejects duplicate mechanic emails.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/CreateMechanicPayload'
      responses:
        201:
          description: Mechanic created
          schema:
            $ref: '#/definitions/Mechanic'
          examples:
            application/json:
              id: 2
              name: John Doe
              email: john.doe@gmail.com
              phone: 123-456-7890
              salary: 25.0
              service_ticket_ids: []
              ticket_count: 0
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        409:
          description: Mechanic with that email already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: A mechanic with that email already exists.
        500:
          description: Failed to create mechanic
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to create mechanic.
    get:
      tags:
        - Mechanics
      summary: List mechanics
      description: Returns a paginated list of mechanics. If pagination values are invalid, all mechanics are returned.
      parameters:
        - in: query
          name: page
          type: integer
          required: false
          default: 1
        - in: query
          name: per_page
          type: integer
          required: false
          default: 10
      responses:
        200:
          description: Mechanic list
          schema:
            type: array
            items:
              $ref: '#/definitions/Mechanic'
          examples:
            application/json:
              - id: 2
                name: test user
                email: test@example.com
                phone: 123-456-7890
                salary: 25.0
                service_ticket_ids: [1]
                ticket_count: 1
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: Login mechanic
      description: Authenticates a mechanic and returns a bearer token and mechanic profile data.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/LoginCredentials'
      responses:
        200:
          description: Login success
          schema:
            $ref: '#/definitions/MechanicLoginResponse'
          examples:
            application/json:
              token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              mechanic:
                id: 2
                name: test user
                email: test@example.com
                phone: 123-456-7890
                salary: 25.0
                service_ticket_ids: [1]
                ticket_count: 1
        400:
          description: Validation error.
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        401:
          description: Invalid email or password.
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Invalid email or password.
...
  /mechanics/{id}:
    put:
      tags:
        - Mechanics
      summary: Update a mechanic
      description: Updates an existing mechanic by ID. Requires a valid mechanic bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UpdateMechanicPayload'
      responses:
        200:
          description: Mechanic updated
          schema:
            $ref: '#/definitions/Mechanic'
          examples:
            application/json:
              id: 2
              name: Updated Name
              email: test@example.com
              phone: 987-654-3210
              salary: 25.0
              service_ticket_ids: [1]
              ticket_count: 1
        400:
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Validation failed for one or more fields.
        404:
          description: Mechanic not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Mechanic not found.
        409:
          description: Mechanic with that email already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: A mechanic with that email already exists.
        500:
          description: Failed to update mechanic
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to update mechanic.
    delete:
      tags:
        - Mechanics
      summary: Delete a mechanic
      description: Deletes a mechanic by ID. Requires a valid mechanic bearer token.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          type: integer
      responses:
        200:
          description: Mechanic deleted
          schema:
            $ref: '#/definitions/MessageResponse'
          examples:
            application/json:
              message: Mechanic 1 deleted.
        404:
          description: Mechanic not found
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Mechanic not found.
        500:
          description: Failed to delete mechanic
          schema:
            $ref: '#/definitions/ErrorResponse'
          examples:
            application/json:
              error: Failed to delete mechanic.
...
definitions:
  LoginCredentials:
    type: object
    properties:
      email:
        type: string
        example: test@example.com
      password:
        type: string
        example: password
    required:
      - email
      - password
...
  Mechanic:
    type: object
    properties:
      id:
        type: integer
        example: 2
      name:
        type: string
        example: test user
      email:
        type: string
        example: test@example.com
      phone:
        type: string
        example: 123-456-7890
      salary:
        type: number
        format: float
        example: 25.0
      service_ticket_ids:
        type: array
        items:
          type: integer
        example: [2, 5]
      ticket_count:
        type: integer
        example: 2
...
  CreateMechanicPayload:
    type: object
    properties:
      name:
        type: string
        example: John Doe
      email:
        type: string
        example: john.doe@gmail.com
      phone:
        type: string
        example: 123-456-7890
      salary:
        type: number
        format: float
        example: 25.0
      password:
        type: string
        example: securepassword
    required:
      - name
      - email
      - phone
      - salary
  UpdateMechanicPayload:
    type: object
    properties:
      name:
        type: string
        example: Updated Name
      email:
        type: string
        example: test@example.com
      phone:
        type: string
        example: 987-654-3210
      salary:
        type: number
        format: float
        example: 25.0
      password:
        type: string
        example: password
...
  MechanicLoginResponse:
    type: object
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      mechanic:
        $ref: '#/definitions/Mechanic'
...
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: Resource deleted successfully.
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: Validation failed for one or more fields.